Atualização para o sabado de integração
Atualização para o sabado de integração

+++ model/area.php
@@ -19,11 +19,28 @@ function getArea($db, $order, $limit) {
 
 // ==================== DELETE ====================
 function delArea($db, $del) {
-    if(odbc_exec($db, 
-        "DELETE FROM area
-        WHERE codArea = $del"
-    )) {
-        return true;
+    $query =    odbc_exec($db, 
+                    "SELECT codAssunto, codArea
+                    FROM assunto"
+                );
+    while($result = odbc_fetch_array($query)) {
+        $assuntos[$result['codAssunto']] = $result['codArea'];
+    }
+    $validaDel = true;
+    foreach ($assuntos as $key => $value) {
+        if($assuntos[$key] = $del) {
+            $validaDel = false;
+        }
+    }
+    if($validaDel) {
+        if(odbc_exec($db, 
+            "DELETE FROM area
+            WHERE codArea = $del"
+        )) {
+            return true;
+        } else {
+            return false;
+        }
     } else {
         return false;
     }
